Add CompositeServices and rename simple_services to services
