+++ experimental/proto-checking/build.sh
@@ -2,7 +2,7 @@
 # protoc --cpp_out=cpp \
 #        --js_out=js \
 #        --python_out=python \
-#        simple_services.proto
+#        services.proto
 
 # Compile gRPC skeletons to Python
-python -m grpc_tools.protoc -I. --python_out=python --grpc_python_out=python simple_services.proto
+python -m grpc_tools.protoc -I. --python_out=python --grpc_python_out=python services.proto
