+++ experimental/proto-checking/python/services_client.py
@@ -1,23 +1,23 @@
 import grpc
-import simple_services_pb2
-import simple_services_pb2_grpc
+import services_pb2
+import services_pb2_grpc
 
 def run_client():
     channel = grpc.insecure_channel('localhost:50051')  # Set the server address and port
-    stub = simple_services_pb2_grpc.SimpleServicesStub(channel)
+    stub = services_pb2_grpc.SimpleServicesStub(channel)
 
     # Test foo
-    foo_request = simple_services_pb2.FooIn(foo_bool=True)
+    foo_request = services_pb2.FooIn(foo_bool=True)
     foo_response = stub.foo_rpc(foo_request)
     print("Foo response: {}".format(foo_response.foo_int))
 
     # Test bar
-    bar_request = simple_services_pb2.BarIn(bar_int=42)
+    bar_request = services_pb2.BarIn(bar_int=42)
     bar_response = stub.bar_rpc(bar_request)
     print("Bar response: \"{}\"".format(bar_response.bar_str))
 
     # Test baz
-    baz_request = simple_services_pb2.BazIn(baz_str="42")
+    baz_request = services_pb2.BazIn(baz_str="42")
     baz_response = stub.baz_rpc(baz_request)
     print("Baz response:", baz_response.baz_bool)
